backend/pyproject.toml
[project]
name = "rag-agent-api"
version = "0.1.0"
description = "FastAPI backend for Physical AI & Robotics textbook Q&A with OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Existing dependencies
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.0",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    # New for Spec-3: RAG Agent API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openai-agents>=0.2.0",
    "openai>=1.60.0",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "pydantic>=2.10.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
